Fixes a conflict between TinyMCE/Spry.
This issue would cause a TinyMCE-enabled textarea field to fail valiation
the first time a user tries to submit an otherwise-valid value in Internet
Explorer (version list unknown).

When TinyMCE injects itself in to the page, it removes the original textarea tag
and replaces it with a hidden form field, preserving the name attribute. It seems
that IE was improperly reporting this attribute as being readonly/disabled the
first time ValiationTextarea asked - I am not sure if Spry later goes on to 'correct'
this or if IE naturally begins reporting it correctly after an aborted submit.

Forcing TinyMCE to dump the current value of the rich text editor in to the hidden
form field fixes the issue. I have put this directly in to the Spry widget to minimize
the changes that need be made.