English inflections fixed up a bit.
The inflector saw some significant field testing. Several issues were noted and fixed:

== Meat ==
The word meat is a mass noun. There is one circumstance in which pluralizing a mass
noun is considered correct, and that is when you are refering to many different
substances. Example - the butcher saw increased profit on some of his meats.

Given the context for the below items, we can say that these are the same substance,
and therefore pluralizing meat is wrong. I will add the word as an uncountable in
the inflector and hope that we never have an item which requires the word meats.

    1. 5 Rotting Meats
    2. 5 Iced Meats
    3. 5 Freshly Hunted Meats

Cheese and fruit have had a similar issue addressed.

== Of ==
These look to be rather straightforward. The phrase should be checked for the
word 'of'. The last word prior to that should be what is inflected, not the
last word in the phrase.

    1. 5 Bar of Golds
    2. 5 Scales of the Great Steel-Snakes
    3. Piece of Woods

== Potato ==
Potato was being pluralized as potatos. It will now show as potatoes.